feat: Controller 별 api 위치 변경
