+++ src/main/java/roomescape/user/controller/AuthController.java
@@ -3,7 +3,6 @@ package roomescape.user.controller;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
-import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CookieValue;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -11,23 +10,21 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.user.dto.request.LoginRequest;
-import roomescape.user.dto.response.BriefUserResponse;
 import roomescape.user.dto.response.LoginCheckResponse;
-import roomescape.user.domain.UserPrinciple;
-import roomescape.user.service.LoginService;
+import roomescape.user.service.UserService;
 
 @RestController
-public class LoginController {
+public class AuthController {
 
-    private final LoginService loginService;
+    private final UserService userService;
 
-    public LoginController(LoginService loginService) {
-        this.loginService = loginService;
+    public AuthController(UserService userService) {
+        this.userService = userService;
     }
 
     @PostMapping("/login")
     public void login(@RequestBody @Valid LoginRequest loginRequest, HttpServletResponse response) {
-        String accessToken = loginService.login(loginRequest);
+        String accessToken = userService.login(loginRequest);
         Cookie cookie = new Cookie("token", accessToken);
         cookie.setHttpOnly(true);
         cookie.setPath("/");
@@ -43,17 +40,6 @@ public class LoginController {
 
     @GetMapping("/login/check")
     public ResponseEntity<LoginCheckResponse> checkLogin(@CookieValue(name = "token") String token) {
-        return ResponseEntity.ok(loginService.checkUserByToken(token));
-    }
-
-    @GetMapping("/login/check-v2")
-    public ResponseEntity<String> test(UserPrinciple userPrinciple) {
-        return ResponseEntity.ok(userPrinciple.name() + " " + userPrinciple.email());
-    }
-
-    @GetMapping("/users/all")
-    public ResponseEntity<List<BriefUserResponse>> getAllUsers() {
-        List<BriefUserResponse> users = loginService.getAllUsers();
-        return ResponseEntity.ok(users);
+        return ResponseEntity.ok(userService.checkUserByToken(token));
     }
 }
