+++ src/main/resources/static/js/reservation-with-member.js
@@ -1,8 +1,8 @@
 let isEditing = false;
-const RESERVATION_API_ENDPOINT = '/reservations';
+const RESERVATION_API_ENDPOINT = '/admin/reservations';
 const TIME_API_ENDPOINT = '/times/all';
 const THEME_API_ENDPOINT = '/themes/all';
-const MEMBER_API_ENDPOINT = '/users/all';
+const MEMBER_API_ENDPOINT = '/admin/users/all';
 const timesOptions = [];
 const themesOptions = [];
 const membersOptions = [];
